requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
enum34==1.1.6
Flask==1.1.1
idna==2.8
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycparser==2.19
pyOpenSSL==19.0.0
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
simple-salesforce==0.74.3
six==1.12.0
typing==3.7.4
urllib3==1.25.3
Werkzeug==0.15.5
